refactor JPA submodels
